test(gateway/MAG-19): cover Supercharger battery display
Change: RM31-gateway-show-session-battery-pct

Tasks: 3.1, 3.2, 4.1

Roadmap: RM31-supercharger-session-verification / tier 4

Ticket: MAG-19

Pipeline-Step: wave-3

+++ internal/gateway/handlers/supercharger.go
@@ -384,7 +384,7 @@ func buildSuperchargerRows(sessions []charging.Session) []fragments.Supercharger
 	return rows
 }
 
-func formatBatteryPct(pct *int16) string {
+func formatBatteryPct(pct *int) string {
 	if pct == nil {
 		return "—"
 	}
